don't clobber r15
The quick_sort function uses r15 to store the original last index, but
the partition macro also uses it for the pivot value, which messes up the
range for the sort, resulting in a bunch of values becoming zero.

This patch makes the partition macro use `r11` instead, which is unused
in this context.